+++ Hearthstone Deck Tracker/Hearthstone/RelatedCardsSystem/Cards/Mage/DiscoAtTheEndOfTime.cs
@@ -1,0 +1,16 @@
+﻿using System.Linq;
+using HearthDb.Enums;
+
+namespace Hearthstone_Deck_Tracker.Hearthstone.RelatedCardsSystem.Cards.Mage;
+
+public class DiscoAtTheEndOfTime : ICardGenerator
+{
+	public string GetCardId() => HearthDb.CardIds.Collectible.Mage.DiscoAtTheEndOfTime;
+
+	public bool IsInGeneratorPool(Card card, GameType gameMode, FormatType format)
+	{
+		return card.GetTag(GameTag.SECRET) > 0 &&
+		       (Helper.WildOnlySets.Contains(card.Set) ||
+		        Helper.ClassicOnlySets.Contains(card.Set));
+	}
+}