+++ Hearthstone Deck Tracker/Hearthstone/RelatedCardsSystem/Cards/Mage/TearReality.cs
@@ -1,0 +1,16 @@
+﻿using System.Linq;
+using HearthDb.Enums;
+
+namespace Hearthstone_Deck_Tracker.Hearthstone.RelatedCardsSystem.Cards.Mage;
+
+public class TearReality : ICardGenerator
+{
+	public string GetCardId() => HearthDb.CardIds.Collectible.Mage.TearReality;
+
+	public bool IsInGeneratorPool(Card card, GameType gameMode, FormatType format)
+	{
+		return card.Type == "Spell" && card.IsClass("Mage") &&
+			(Helper.WildOnlySets.Contains(card.Set) ||
+		        Helper.ClassicOnlySets.Contains(card.Set));
+	}
+}