+++ Hearthstone Deck Tracker/Hearthstone/RelatedCardsSystem/Cards/Warlock/SketchArtist.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 
 namespace Hearthstone_Deck_Tracker.Hearthstone.RelatedCardsSystem.Cards.Warlock;
 
-public class SketchArtist : ICardWithHighlight, ISpellSchoolTutor
+public class SketchArtist : ICardWithHighlight, ISpellSchoolTutor, ICardGenerator
 {
 	public string GetCardId() => HearthDb.CardIds.Collectible.Warlock.SketchArtist;
 
@@ -11,5 +11,12 @@ public class SketchArtist : ICardWithHighlight, ISpellSchoolTutor
 		HighlightColorHelper.GetHighlightColor(card.GetTag(GameTag.SPELL_SCHOOL) == (int)SpellSchool.SHADOW);
 
 	public int[] TutoredSpellSchools { get; } = { (int)SpellSchool.SHADOW };
+
+	public bool IsInGeneratorPool(Card card, GameType gameMode, FormatType format)
+	{
+		return card.Type == "Spell" &&
+		       card.GetTag(GameTag.SPELL_SCHOOL) == (int)SpellSchool.SHADOW &&
+		       card.IsCardLegal(gameMode, format);
+	}
 
 }