+++ Hearthstone Deck Tracker/Hearthstone/RelatedCardsSystem/ICardGenerator.cs
@@ -1,0 +1,8 @@
+﻿using HearthDb.Enums;
+
+namespace Hearthstone_Deck_Tracker.Hearthstone.RelatedCardsSystem;
+
+public interface ICardGenerator : ICard
+{
+	bool IsInGeneratorPool(Card card, GameType gameMode, FormatType format);
+}