+++ Hearthstone Deck Tracker/Hearthstone/RelatedCardsSystem/ICardWithHighlight.cs
@@ -7,8 +7,3 @@ public interface ICardWithHighlight : ICard
 {
 	HighlightColor ShouldHighlight(Card card, IEnumerable<Card> deck);
 }
-
-public interface ISpellSchoolTutor : ICardWithHighlight
-{
-	int[] TutoredSpellSchools { get; }
-}