+++ Hearthstone Deck Tracker/Hearthstone/RelatedCardsSystem/ISpellSchoolTutor.cs
@@ -1,0 +1,6 @@
+﻿namespace Hearthstone_Deck_Tracker.Hearthstone.RelatedCardsSystem;
+
+public interface ISpellSchoolTutor : ICardWithHighlight
+{
+	int[] TutoredSpellSchools { get; }
+}