+++ Hearthstone Deck Tracker/Hearthstone/RelatedCardsSystem/RelatedCardsManager.cs
@@ -13,42 +13,56 @@ public class RelatedCardsManager
 	private Dictionary<string, ICardWithRelatedCards>? _relatedCards;
 	private Dictionary<string, ICardWithHighlight>? _highlightCards;
 	private Dictionary<string, ISpellSchoolTutor>? _spellSchoolTutorCards;
+	private Dictionary<string, ICardGenerator>? _cardGeneratorCards;
 
 	public Dictionary<string, ICardWithRelatedCards> RelatedCards => _relatedCards ??= InitializeRelatedCards();
 	public Dictionary<string, ICardWithHighlight> HighlightCards  => _highlightCards ??= InitializeHighlightCards();
-
 	public Dictionary<string, ISpellSchoolTutor> SpellSchoolTutorCards  => _spellSchoolTutorCards ??= InitializeSpellSchoolTutorCards();
+	public Dictionary<string, ICardGenerator> CardGeneratorCards  => _cardGeneratorCards ??= InitializeCardGeneratorCards();
 
 
 	private Dictionary<string, ICardWithRelatedCards> InitializeRelatedCards()
 	{
-		var (relatedCardsDict, highlightCardsDict, spellSchoolTutorCardsDict ) = InitializeCards();
+		var (relatedCardsDict, highlightCardsDict, spellSchoolTutorCardsDict, generatorsDict ) = InitializeCards();
 		_highlightCards = highlightCardsDict;
 		_spellSchoolTutorCards = spellSchoolTutorCardsDict;
+		_cardGeneratorCards = generatorsDict;
 		return relatedCardsDict;
 	}
 
 	private Dictionary<string, ICardWithHighlight> InitializeHighlightCards()
 	{
-		var (relatedCardsDict, highlightCardsDict, spellSchoolTutorCardsDict ) = InitializeCards();
+		var (relatedCardsDict, highlightCardsDict, spellSchoolTutorCardsDict, generatorsDict ) = InitializeCards();
 		_relatedCards = relatedCardsDict;
 		_spellSchoolTutorCards = spellSchoolTutorCardsDict;
+		_cardGeneratorCards = generatorsDict;
 		return highlightCardsDict;
 	}
 
 	private Dictionary<string, ISpellSchoolTutor> InitializeSpellSchoolTutorCards()
 	{
-		var (relatedCardsDict, highlightCardsDict, spellSchoolTutorCardsDict ) = InitializeCards();
+		var (relatedCardsDict, highlightCardsDict, spellSchoolTutorCardsDict, generatorsDict ) = InitializeCards();
 		_relatedCards = relatedCardsDict;
 		_highlightCards = highlightCardsDict;
+		_cardGeneratorCards = generatorsDict;
 		return spellSchoolTutorCardsDict;
+	}
+
+	private Dictionary<string, ICardGenerator> InitializeCardGeneratorCards()
+	{
+		var (relatedCardsDict, highlightCardsDict, spellSchoolTutorCardsDict, generatorsDict ) = InitializeCards();
+		_relatedCards = relatedCardsDict;
+		_highlightCards = highlightCardsDict;
+		_spellSchoolTutorCards = spellSchoolTutorCardsDict;
+		return generatorsDict;
 	}
 
 	private (
 		Dictionary<string, ICardWithRelatedCards>,
 		Dictionary<string, ICardWithHighlight>,
-		Dictionary<string, ISpellSchoolTutor>
-		) InitializeCards()
+		Dictionary<string, ISpellSchoolTutor>,
+		Dictionary<string, ICardGenerator> )
+		InitializeCards()
 	{
 		var cards = Assembly.GetAssembly(typeof(ICard)).GetTypes()
 			.Where(t => t.IsClass && !t.IsAbstract && typeof(ICard).IsAssignableFrom(t));
@@ -56,6 +70,7 @@ public class RelatedCardsManager
 		var relatedCardsDict = new Dictionary<string, ICardWithRelatedCards>();
 		var highlightCardsDict = new Dictionary<string, ICardWithHighlight>();
 		var spellSchoolTutorCardsDict = new Dictionary<string, ISpellSchoolTutor>();
+		var cardGeneratorCardsDict = new Dictionary<string, ICardGenerator>();
 
 
 		foreach(var card in cards)
@@ -76,9 +91,14 @@ public class RelatedCardsManager
 			{
 				spellSchoolTutorCardsDict[tutor.GetCardId()] = tutor;
 			}
+
+			if(cardInstance is ICardGenerator generator)
+			{
+				cardGeneratorCardsDict[generator.GetCardId()] = generator;
+			}
 		}
 
-		return (relatedCardsDict, highlightCardsDict, spellSchoolTutorCardsDict);
+		return (relatedCardsDict, highlightCardsDict, spellSchoolTutorCardsDict, cardGeneratorCardsDict);
 	}
 
 	public ICardWithHighlight? GetCardWithHighlight(string cardId)